Re-designed proposal "Reason" to allow unlimited text.
Now, instead of limiting the player to a certain amount of characters, they can type for as long as they want. The full reason may not always be displayed during the proposal announcement, but now anyone can view the full reason by hovering over the text. I like this a lot. : )